+++ src/lib.rs
@@ -1,6 +1,3 @@
-#[macro_use]
-extern crate num_derive;
-
 pub use context::HeifContext;
 pub use encoder::{EncoderParametersTypes, EncodingOptions};
 pub use enums::*;